requirements.txt
joblib
matplotlib
numpy
pandas
pillow
torchinfo
tqdm
typing-extensions

--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
